Stripe fully implemented
This is going to require alot of installation and steps for it to work so follow the steps stated below :

1. Cd to your react native project
2. Type ‘Firebase login’ (then key in your login password)
3. Npm install -g firebase-tools
4. (pick our aspproject on firebase)
5. Pick javascript
6. Pick yes to install dependencies
(You should see a functions folder created in your react native project, within there will be an index.js where the firebase functions are ran on.)

7.  Within the functions folder, type ‘npm install stripe’ to install stripe dependencies

8. Replace the index.js in this function folder with the one in the Firebase indexfile

9. Now go back to the main project folder on directory:
Type in ‘Npm install axios’ - this will facilitate the request data to the index.js at the functions folder.

10. On firebase website itself - the mode should be on blaze mode (pay on the go).

after completing all the above stepsit should work.